chain select strong refactor
